refactor client and fix tests
