+++ client/arguments.serverless
@@ -1,0 +1,1 @@
+{"circuit": {"__type__": "QuantumCircuit", "__value__": "eJwL9Az29gzhZmRiYYAAxtRCBu40Bg4gEyYEAkwgKSibJTmzKLk0s0TX0MwUJl9dW8gI1sLIWMiAChiRjAABZrgMB8xABuwgyNk4wj2xJBVkIsRYEMFUCDaEDWo0Iy7NcK0MaTD7ErW9ZUzYuB3AmlmQ3ITmXEbnUBR7meEm7Nhj+TtzrYgDXMBy/hb951zf7OECmbve2Z8KZUCo2O649MOVYiYHBla0AEEHcAeDXPUfCcAUAADnyDSq"}, "array": {"__type__": "ndarray", "__value__": "eJyb7BfqGxDJyFDGUK2eklqcXKRupaBuk2ahrqOgnpZfVFKUmBefX5SSChJ3S8wpTgWKF2ckFqQC+RpGOgqGmjoKtQpkAy4GMHB1YIACABOFG/M="}}
