+++ client/qiskit_serverless/core/clients/RayClient.py
@@ -26,78 +26,60 @@ Qiskit Serverless provider
     ComputeResource
     ServerlessClient
 """
-# pylint: disable=duplicate-code
-from abc import ABC, abstractmethod
+import json
+import warnings
 from typing import Optional, List, Dict, Any, Union
+from uuid import uuid4
 
+from ray.dashboard.modules.job.sdk import JobSubmissionClient
 from qiskit_ibm_runtime import QiskitRuntimeService
 
-from qiskit_serverless.core.job import Job
+from qiskit_serverless.core.constants import (
+    OT_PROGRAM_NAME,
+    ENV_JOB_ARGUMENTS,
+)
+from qiskit_serverless.core.job import (
+    Job,
+)
 from qiskit_serverless.core.function import QiskitFunction
-from qiskit_serverless.utils import JsonSerializable
-from qiskit_serverless.visualizaiton import Widget
+from qiskit_serverless.serializers.program_serializers import (
+    QiskitObjectsEncoder,
+)
 
 TIMEOUT = 30
 
+from qiskit_serverless.core.client import BaseClient
 
-class BaseClient(JsonSerializable, ABC):
-    """
-    A client class for specifying custom compute resources.
 
-    Example:
-        >>> client = BaseClient(
-        >>>    name="<NAME>",
-        >>>    host="<HOST>",
-        >>>    token="<TOKEN>",
-        >>>    compute_resource=ComputeResource(
-        >>>        name="<COMPUTE_RESOURCE_NAME>",
-        >>>        host="<COMPUTE_RESOURCE_HOST>"
-        >>>    ),
-        >>> )
-    """
+class RayClient(BaseClient):
+    """RayClient."""
 
-    def __init__(  # pylint:  disable=too-many-positional-arguments
-        self, name: str, host: Optional[str] = None, token: Optional[str] = None
-    ):
-        """
-        Initialize a BaseClient instance.
+    def __init__(self, host: str):
+        """Ray client
 
         Args:
-            name: name of client
-            host: host of client a.k.a managers host
-            token: authentication token for manager
+            host: ray head node host
+
+        Example:
+            >>> ray_provider = RayClient("http://localhost:8265")
         """
-        self.name = name
-        self.host = host
-        self.token = token
-
-    # ?¿¿?¿?
-    @classmethod
-    def from_dict(cls, dictionary: dict):
-        return BaseClient(**dictionary)
-
-    def __eq__(self, other):
-        if isinstance(other, BaseClient):
-            return self.name == other.name
-
-        return False
-
-    def __repr__(self):
-        return f"<{self.name}>"
+        super().__init__("ray-client", host)
+        self.jobSubmissionClient = JobSubmissionClient(host)
 
     ####################
     ####### JOBS #######
     ####################
-    @abstractmethod
+
     def get_jobs(self, **kwargs) -> List[Job]:
         """Return list of jobs.
 
         Returns:
             list of jobs.
         """
-        pass
+        return [
+            Job(job.job_id, client=self) for job in self.jobSubmissionClient.list_jobs()
+        ]
 
-    @abstractmethod
     def get_job(self, job_id: str) -> Optional[Job]:
         """Returns job by job id.
 
@@ -107,93 +89,91 @@ class BaseClient(JsonSerializable, ABC):
         Returns:
             Job instance
         """
-        pass
+        return Job(
+            self.jobSubmissionClient.get_job_info(job_id).submission_id, client=self
+        )
 
-    @abstractmethod
     def run(
         self,
         program: Union[QiskitFunction, str],
         arguments: Optional[Dict[str, Any]] = None,
     ) -> Job:
-        """Execute a program as a async job.
+        if not isinstance(program, QiskitFunction):
+            warnings.warn(
+                "`run` doesn't support program str yet. "
+                "Send a QiskitFunction instead. "
+            )
+            return NotImplementedError
 
-        Example:
-            >>> serverless = QiskitServerless()
-            >>> program = QiskitFunction(
-            >>>     "job.py",
-            >>>     arguments={"arg1": "val1"},
-            >>>     dependencies=["requests"]
-            >>> )
-            >>> job = serverless.run(program)
-            >>> # <Job | ...>
+        arguments = arguments or {}
+        entrypoint = f"python {program.entrypoint}"
 
-        Args:
-            arguments: arguments to run program with
-            program: Program object
+        # set program name so OT can use it as parent span name
+        env_vars = {
+            **(program.env_vars or {}),
+            **{OT_PROGRAM_NAME: program.title},
+            **{ENV_JOB_ARGUMENTS: json.dumps(arguments, cls=QiskitObjectsEncoder)},
+        }
 
-        Returns:
-            Job
-        """
-        pass
+        job_id = self.jobSubmissionClient.submit_job(
+            entrypoint=entrypoint,
+            submission_id=f"qs_{uuid4()}",
+            runtime_env={
+                "working_dir": program.working_dir,
+                "pip": program.dependencies,
+                "env_vars": env_vars,
+            },
+        )
+        return Job(job_id=job_id, client=self)
 
-    @abstractmethod
     def status(self, job_id: str) -> str:
         """Check status."""
-        pass
+        return self.jobSubmissionClient.get_job_status(job_id).value
 
-    @abstractmethod
     def stop(
         self, job_id: str, service: Optional[QiskitRuntimeService] = None
     ) -> Union[str, bool]:
         """Stops job/program."""
-        pass
+        return self.jobSubmissionClient.stop_job(job_id)
 
-    @abstractmethod
     def result(self, job_id: str) -> Any:
         """Return results."""
-        pass
+        return self.logs(job_id)
 
-    @abstractmethod
     def logs(self, job_id: str) -> str:
         """Return logs."""
-        pass
+        return self.jobSubmissionClient.get_job_logs(job_id)
 
-    @abstractmethod
     def filtered_logs(self, job_id: str, **kwargs) -> str:
         """Return filtered logs."""
-        pass
+        raise NotImplementedError
 
     #########################
     ####### Functions #######
     #########################
 
-    @abstractmethod
     def upload(self, program: QiskitFunction) -> Optional[QiskitFunction]:
         """Uploads program."""
-        pass
+        raise NotImplementedError("Upload is not available for RayClient.")
 
-    @abstractmethod
     def get_programs(self, **kwargs) -> List[QiskitFunction]:
         """Returns list of available programs."""
-        pass
+        raise NotImplementedError("get_programs is not available for RayClient.")
 
-    @abstractmethod
     def get_program(
         self, title: str, provider: Optional[str] = None
     ) -> Optional[QiskitFunction]:
         """Returns program based on parameters."""
-        pass
+        raise NotImplementedError("get_program is not available for RayClient.")
 
     #####################
     ####### FILES #######
     #####################
 
-    @abstractmethod
     def files(self) -> List[str]:
         """Returns list of available files produced by programs to download."""
-        pass
+        raise NotImplementedError("files is not available for RayClient.")
 
-    @abstractmethod
     def file_download(
         self,
         file: str,
@@ -201,22 +181,12 @@ class BaseClient(JsonSerializable, ABC):
         download_location: str = "./",
     ):
         """Download file."""
-        pass
+        raise NotImplementedError("file_download is not available for RayClient.")
 
-    @abstractmethod
     def file_delete(self, file: str):
         """Deletes file uploaded or produced by the programs,"""
-        pass
+        raise NotImplementedError("file_delete is not available for RayClient.")
 
-    @abstractmethod
     def file_upload(self, file: str):
         """Upload file."""
-        pass
-
-    #####################
-    ####### ????? #######
-    #####################
-
-    def widget(self):
-        """Widget for information about provider and jobs."""
-        return Widget(self).show()
+        raise NotImplementedError("file_upload is not available for RayClient.")
