+++ client/tests/core/test_pattern.py
@@ -1,13 +1,15 @@
 """Tests jobs."""
 import os
 
+import ray
+
 from testcontainers.compose import DockerCompose
 
-from qiskit_serverless import BaseClient
-from qiskit_serverless.core import ComputeResource
+from qiskit_serverless import RayClient, QiskitFunction
 from qiskit_serverless.core.job import Job
-from qiskit_serverless.core.function import QiskitFunction
-from tests.utils import wait_for_job_client, wait_for_job_completion
+from tests.utils import wait_for_ray_ready, wait_for_job_completion
+
+import time
 
 resources_path = os.path.join(
     os.path.dirname(os.path.abspath(__file__)), "../resources"
@@ -22,14 +24,11 @@ def test_program():
     ) as compose:
         host = compose.get_service_host("testrayhead", 8265)
         port = compose.get_service_port("testrayhead", 8265)
+        connection_url = f"http://{host}:{port}"
 
-        serverless = BaseClient(
-            name="docker",
-            compute_resource=ComputeResource(
-                name="docker", host=host, port_job_server=port
-            ),
-        )
-        wait_for_job_client(serverless)
+        wait_for_ray_ready(connection_url)
+
+        serverless = RayClient(host=connection_url)
 
         program = QiskitFunction(
             title="simple_job",
@@ -49,7 +48,7 @@ def test_program():
         assert job.in_terminal_state()
         assert job.status() == "DONE"
 
-        recovered_job = serverless.get_job_by_id(job.job_id)
+        recovered_job = serverless.get_job(job.job_id)
         assert recovered_job.job_id == job.job_id
         assert "42" in recovered_job.logs()
         assert recovered_job.in_terminal_state()
